Add stop streaming button
