+++ llm_studio/app_utils/sections/chat.py
@@ -1,28 +1,17 @@
-import asyncio
 import gc
 import logging
 import os
-import threading
-from typing import Any, Callable, Dict, List, Optional
 
-import numpy as np
 import torch
 from accelerate import dispatch_model, infer_auto_device_map
 from accelerate.utils import get_balanced_memory
 from h2o_wave import Q
 from h2o_wave import data as chat_data
 from h2o_wave import ui
-from transformers import AutoTokenizer, TextStreamer
 
-from llm_studio.app_utils.utils import (
-    get_experiments,
-    get_ui_elements,
-    parse_ui_elements,
-    set_env,
-)
+from llm_studio.app_utils.utils import get_experiments, get_ui_elements, set_env
 from llm_studio.python_configs.base import DefaultConfigProblemBase
 from llm_studio.src.datasets.text_utils import get_tokenizer
-from llm_studio.src.models.text_causal_language_modeling_model import Model
 from llm_studio.src.utils.config_utils import (
     NON_GENERATION_PROBLEM_TYPES,
     load_config_yaml,
@@ -30,54 +19,6 @@ from llm_studio.src.utils.config_utils import (
 from llm_studio.src.utils.modeling_utils import load_checkpoint
 
 logger = logging.getLogger(__name__)
-
-USER = True
-BOT = False
-
-
-class WaveChatStreamer(TextStreamer):
-    def __init__(
-        self,
-        tokenizer: AutoTokenizer,
-        q: Q,
-        text_cleaner: Optional[Callable] = None,
-        **decode_kwargs,
-    ):
-        """
-        Updates the chabot card in a streaming fashion
-        """
-        super().__init__(tokenizer, skip_prompt=True, **decode_kwargs)
-        self.text_cleaner = text_cleaner
-        self.words_predicted_answer: List[str] = []
-        self.q = q
-        self.loop = asyncio.get_event_loop()
-        self.finished = False
-
-    def on_finalized_text(self, text: str, stream_end: bool = False):
-        self.words_predicted_answer += [text]
-        self.loop.create_task(self.push_to_chat())
-
-    async def push_to_chat(self):
-        self.q.page["experiment/display/chat"].data[-1] = [self.answer, BOT]
-        await self.q.page.save()
-
-    @property
-    def answer(self):
-        """
-        Create the answer by joining the generated words.
-        By this, self.text_cleaner does not need to be idempotent.
-        """
-        answer = "".join(self.words_predicted_answer)
-        if answer.endswith(self.tokenizer.eos_token):
-            # text generation is stopped
-            answer = answer.replace(self.tokenizer.eos_token, "")
-        if self.text_cleaner:
-            answer = self.text_cleaner(answer)
-        return answer
-
-    def end(self):
-        super().end()
-        self.finished = True
 
 
 async def chat_tab(q: Q, load_model=True):
@@ -138,7 +79,7 @@ async def chat_tab(q: Q, load_model=True):
         data=chat_data(fields="content from_user", t="list"),  # type: ignore
         name="experiment/display/chat/chatbot",
     )
-    q.page["experiment/display/chat"].data += [initial_message, BOT]
+    q.page["experiment/display/chat"].data += [initial_message, False]
 
     option_items = get_ui_elements(
         cfg=q.client["experiment/display/chat/cfg"].prediction,
@@ -148,10 +89,19 @@ async def chat_tab(q: Q, load_model=True):
     q.page["experiment/display/chat/settings"] = ui.form_card(
         box="second",
         items=[
-            ui.button(
-                name="experiment/display/chat/clear_history",
-                label="Clear History",
-                primary=True,
+            ui.buttons(
+                [
+                    ui.button(
+                        name="experiment/display/chat/clear_history",
+                        label="Clear History",
+                        primary=True,
+                    ),
+                    ui.button(
+                        name="experiment/display/chat/abort_stream",
+                        label="Stop Streaming",
+                        primary=True,
+                    ),
+                ]
             ),
             ui.expander(
                 name="chat_settings",
@@ -216,110 +166,6 @@ def gpu_is_blocked(q, gpu_id):
     return gpu_blocked
 
 
-@torch.inference_mode(mode=True)
-async def chat_update(q: Q) -> None:
-    """
-    Update the chatbot with the new message.
-    """
-    q.client["experiment/display/chat/finished"] = False
-
-    cfg_prediction = parse_ui_elements(
-        cfg=q.client["experiment/display/chat/cfg"].prediction,
-        q=q,
-        pre="chat/cfg_predictions/cfg/",
-    )
-    logger.info(f"Using chatbot config: {cfg_prediction}")
-    q.client["experiment/display/chat/cfg"].prediction = cfg_prediction
-
-    prompt = q.client["experiment/display/chat/chatbot"]
-
-    message = [prompt, USER]
-    q.client["experiment/display/chat/messages"].append(message)
-    q.page["experiment/display/chat"].data += message
-    q.page["experiment/display/chat"].data += ["", BOT]
-    await q.page.save()
-
-    cfg = q.client["experiment/display/chat/cfg"]
-    model: Model = q.client["experiment/display/chat/model"]
-    tokenizer = q.client["experiment/display/chat/tokenizer"]
-
-    full_prompt = ""
-    if len(q.client["experiment/display/chat/messages"]):
-        for prev_message in q.client["experiment/display/chat/messages"][
-            -(cfg.prediction.num_history + 1) :
-        ]:
-            if prev_message[1] is USER:
-                prev_message = cfg.dataset.dataset_class.parse_prompt(
-                    cfg, prev_message[0]
-                )
-            else:
-                prev_message = prev_message[0]
-                if cfg.dataset.add_eos_token_to_answer:
-                    prev_message += cfg._tokenizer_eos_token
-
-            full_prompt += prev_message
-    logger.info(f"Full prompt: {full_prompt}")
-    inputs = cfg.dataset.dataset_class.encode(
-        tokenizer, full_prompt, cfg.tokenizer.max_length_prompt, "left"
-    )
-    inputs["prompt_input_ids"] = (
-        inputs.pop("input_ids").unsqueeze(0).to(cfg.environment._device)
-    )
-    inputs["prompt_attention_mask"] = (
-        inputs.pop("attention_mask").unsqueeze(0).to(cfg.environment._device)
-    )
-
-    def text_cleaner(text: str) -> str:
-        return cfg.dataset.dataset_class.clean_output(
-            output={"predicted_text": np.array([text])}, cfg=cfg
-        )["predicted_text"][0]
-
-    if cfg.prediction.num_beams == 1:
-        streamer = WaveChatStreamer(tokenizer=tokenizer, q=q, text_cleaner=text_cleaner)
-        # Need to start generation in a separate thread, otherwise streaming is blocked
-        thread = threading.Thread(
-            target=generate,
-            kwargs=dict(model=model, inputs=inputs, cfg=cfg, streamer=streamer),
-        )
-        try:
-            thread.start()
-        finally:
-            while True:
-                if streamer.finished:
-                    thread.join()
-                    predicted_text = streamer.answer
-                    break
-                await q.sleep(1)
-    else:
-        # ValueError: `streamer` cannot be used with beam search (yet!).
-        # Make sure that `num_beams` is set to 1.
-        logger.info("Not streaming output, as it cannot be used with beam search.")
-        q.page["experiment/display/chat"].data[-1] = ["...", BOT]
-        await q.page.save()
-        predicted_answer_ids = generate(model, inputs, cfg)[0]
-        predicted_text = tokenizer.decode(
-            predicted_answer_ids, skip_special_tokens=True
-        )
-        predicted_text = text_cleaner(predicted_text)
-
-    logger.info(f"Predicted Answer: {predicted_text}")
-    message = [predicted_text, BOT]
-    q.client["experiment/display/chat/messages"].append(message)
-    q.page["experiment/display/chat"].data[-1] = message
-
-    del inputs
-    gc.collect()
-    torch.cuda.empty_cache()
-
-    q.client["experiment/display/chat/finished"] = True
-
-
-def generate(model: Model, inputs: Dict, cfg: Any, streamer: TextStreamer = None):
-    with torch.cuda.amp.autocast():
-        output = model.generate(batch=inputs, cfg=cfg, streamer=streamer).detach().cpu()
-    return output
-
-
 def load_cfg_model_tokenizer(
     experiment_path: str, merge: bool = False, device: str = "cuda:0"
 ):
@@ -371,15 +217,3 @@ def load_cfg_model_tokenizer(
     model.backbone.use_cache = True
 
     return cfg, model, tokenizer
-
-
-async def show_chat_is_running_dialog(q):
-    q.page["meta"].dialog = ui.dialog(
-        title="Text Generation is streaming.",
-        name="chatbot_running_dialog",
-        items=[
-            ui.text("Please wait till the text generation has stopped."),
-        ],
-        closable=True,
-    )
-    await q.page.save()
